api-server/requirements/base.txt
asgiref==3.2.10
Django==3.1
django-filter==2.4.0
djangorestframework==3.11.1
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
